+++ src/screens/CalenderScreen.jsx
@@ -1,73 +1,9 @@
 import React from 'react';
 import {
-  Text, View, StyleSheet, FlatList,
+  Text, View, StyleSheet, FlatList, TouchableOpacity
 } from 'react-native';
+import { ModalOpen, fakedata } from '../components/CalenderDetail';
 
-const fakedata = [
-  {
-    id: 1,
-    name: 'Calculus',
-    goalId: 3,
-    startTime: new Date(2021, 1, 2, 9, 0),
-    endTime: new Date(2021, 1, 2, 11, 0),
-  },
-  {
-    id: 2,
-    name: 'Codewars',
-    goalId: 2,
-    startTime: new Date(2021, 1, 2, 13, 0),
-    endTime: new Date(2021, 1, 2, 15, 0)
-  },
-  {
-    id: 3,
-    name: 'Trigonmetry',
-    goalId: 3,
-    startTime: new Date(2021, 1, 2, 15, 30),
-    endTime: new Date(2021, 1, 2, 17, 30)
-  },
-  {
-    id: 4,
-    name: 'Calculus',
-    goalId: 3,
-    startTime: new Date(2021, 1, 3, 9, 0),
-    endTime: new Date(2021, 1, 3, 11, 0),
-  },
-  {
-    id: 5,
-    name: 'Codewars',
-    goalId: 2,
-    startTime: new Date(2021, 1, 3, 13, 0),
-    endTime: new Date(2021, 1, 3, 15, 0)
-  },
-  {
-    id: 6,
-    name: 'Trigonmetry',
-    goalId: 3,
-    startTime: new Date(2021, 1, 3, 15, 30),
-    endTime: new Date(2021, 1, 3, 17, 30)
-  },
-  {
-    id: 7,
-    name: 'Calculus',
-    goalId: 3,
-    startTime: new Date(2021, 1, 4, 9, 0),
-    endTime: new Date(2021, 1, 4, 11, 0),
-  },
-  {
-    id: 8,
-    name: 'Codewars',
-    goalId: 2,
-    startTime: new Date(2021, 1, 4, 13, 0),
-    endTime: new Date(2021, 1, 4, 15, 0)
-  },
-  {
-    id: 9,
-    name: 'Trigonmetry',
-    goalId: 3,
-    startTime: new Date(2021, 1, 4, 15, 30),
-    endTime: new Date(2021, 1, 4, 17, 30)
-  },
-];
 
 const monthNames = [ 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
 function addZero(i) {
@@ -77,28 +13,32 @@ function addZero(i) {
   return i;
 };
 
-const CalenderScreen = () => (
-  <FlatList
-    data={fakedata}
-    keyExtractor={(item) => item.id.toString()}
-    renderItem={({ item }) => (
-      <View style={styles.Date_and_Schedule}>
-        <View style={styles.Date}>
-          <Text> { monthNames[ item.startTime.getMonth()] } </Text>
-          <Text style={{fontSize:16}}> { item.startTime.getDate() } </Text>
-        </View>
-        <View style={styles.Schedule}>
-          <View style={styles.Schedule_Name}>
-            <Text style={{fontSize:16}}> { item.name } </Text>
-          </View>
-          <View style={styles.Schedule_Time}>
-            <Text> { addZero(item.startTime.getHours()) }:{ addZero(item.startTime.getMinutes()) } - { addZero(item.endTime.getHours()) }:{ addZero(item.endTime.getMinutes()) } </Text>
-          </View>
-        </View>
-      </View> 
-    )}
-  />
-);
+const CalenderScreen = () => {
+  return(
+    <FlatList
+      data={fakedata}
+      keyExtractor={(item) => item.id.toString()}
+      renderItem={({ item }) => (
+        <TouchableOpacity onPress={() => ModalOpen(1) }>
+          <View style={styles.Date_and_Schedule}>
+            <View style={styles.Date}>
+              <Text> { monthNames[ item.startTime.getMonth()] } </Text>
+              <Text style={{fontSize:16}}> { item.startTime.getDate() } </Text>
+            </View>
+            <View style={styles.Schedule}>
+              <View style={styles.Schedule_Name}>
+                <Text style={{fontSize:16}}> { item.name } </Text>
+              </View>
+              <View style={styles.Schedule_Time}>
+                <Text> { addZero(item.startTime.getHours()) }:{ addZero(item.startTime.getMinutes()) } - { addZero(item.endTime.getHours()) }:{ addZero(item.endTime.getMinutes()) } </Text>
+              </View>
+            </View>
+          </View> 
+        </TouchableOpacity>
+      )}
+    />
+  )
+};
 
 const styles = StyleSheet.create({
   Date_and_Schedule: {
@@ -129,7 +69,7 @@ const styles = StyleSheet.create({
     justifyContent: 'flex-end',
     flexDirection: 'row',
     flex: 1
-  }
+  },
 });
 
 export default CalenderScreen;
